backend/pyproject.toml
[tool.poetry]
name = "storymint"
version = "0.1.0"
description = "School of Solana Project"
authors = ["krill <hey@sol-warrior.com>"]
license = "MIT"
homepage = "https://github.com/heysolwarrior/storymint"
repository = "https://github.com/heysolwarrior/storymint"
keywords = ["django", "storymint"]
classifiers = ["Framework :: Django", "Development Status :: 3 - Alpha", "Operating System :: OS Independent"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django-cors-headers = "*"
django-filter = "*"
djangorestframework = "*"
djangorestframework-simplejwt = "*"
django-semantic-admin = "*"
django-storages = {version = "*", extras = ["google"]}
google-generativeai = "*"
gunicorn = "*"
invoke = "*"
ipython = "*"
psycopg2-binary = "*"
python-decouple = "*"
sentry-sdk = {version = "*", extras = ["django"]}
solathon = "*"
whitenoise = "*"

[tool.poetry.dev-dependencies]
black = "*"
django-debug-toolbar = "*"
ruff = "<0.5.0"

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    "migrations",
]
lint.ignore = [
    "ANN002", # Missing type annotation for *{name}
    "ANN003", # Missing type annotation for **{name}
    "ANN101", # Missing type annotation for {name} in method
    "ANN102", # Missing type annotation for {name} in classmethod
    "ANN401", # Dynamically typed expressions (typing.Any) are disallowed in {name}
    "B007", # Loop control variable not used within loop body
    "B904", # Within an except clause, raise exceptions with...
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D106", # Missing docstring in public nested class, i.e. Meta
    "D203", # 1 blank line required before class docstring
    "D205", # 1 blank line required between summary line and description
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line
    "D400", # First line should end with a period
    "D401", # First line should be in imperative mood
    "D403", # First word of the first line should be properly capitalized
    "D404", # First word of the docstring should not be This
    "D406", # Section name should end with a newline
    "D407", # Missing dashed underline after section
    "D411", # Missing blank line before section
    "D413", # Missing blank line after last section
    "D415", # First line should end with a period, question mark, or exclamation point
    "D417", # Missing argument description in the docstring: {name}
    "E501", # Line too long
    "E721", # Do not compare types, use `isinstance()`
    "N803", # Argument name should be lowercase
    "N806", # Variable in function should be lowercase
    "PTH123", # open() should be replaced by Path.open()
    "UP032", # [*] Use f-string instead of `format` call
]
lint.select = [
    "ANN", # flake8-annotations
    "B", # flake8-bugbear
    "D", # pydocstyle
    "E", # pycodestyle
    "F", # Pyflakes
    "I", # isort
    "INT", # flake8-gettext
    "N", # pep8-naming
    "PTH", # flake8-use-pathlib
    "UP", # pyupgrade
    "RUF100", # unused-noqa
    "RUF200", # invalid-pyproject-toml
    "W", # pycodestyle
]
lint.unfixable = ["F401", "F841"]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["ANN", "D"]

[tool.ruff.lint.isort]
known-first-party = ["storymint"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
